requirements.txt
Flask
google-cloud-vision
google-auth
requests
gunicorn
opencv-python

